Change feature in change_days.py

+++ handlers/delete.py
@@ -2,7 +2,7 @@ from aiogram import Router, F
 from aiogram.fsm.context import FSMContext
 from aiogram.filters.state import StatesGroup, State
 from aiogram import types
-from db.func_for_db import show_all_added_material, delete_one, delete_all, delete_several
+from db.func_for_db import show_all_added_material, delete_one, delete_all, delete_several, get_text_to_repeat
 from keyboards.client_keyboards import kb_for_delete
 from create_bot import bot
 from additional_func import divide
@@ -53,16 +53,26 @@ async def delete_one_handler(callback: types.CallbackQuery, state: FSMContext):
 @router.message(DeleteText.choose_deletion_type)
 async def finish_delete_text_handler(message: types.Message, state: FSMContext):
     deletion_type = await state.get_data()
+    text_to_repeat = [i[0] for i in await get_text_to_repeat(message)]
     if deletion_type['deletion_type'] == 'delete_several':
         phrases_for_deletion = await divide(message.text, ',')
-        await delete_several(phrases_for_deletion, message)
-        await message.answer(Russian.DELETE_SEVERAL_POSITIVE)
-        await state.clear()
-    elif deletion_type['deletion_type'] == 'delete_one':
         try:
+            for text in phrases_for_deletion:
+                if text in text_to_repeat:
+                    continue
+                else:
+                    raise ValueError('Invalid text to delete')
+            await delete_several(phrases_for_deletion, message)
+            await message.answer(Russian.DELETE_SEVERAL_POSITIVE)
+            await state.clear()
+        except:
+            await message.answer(Russian.DELETE_SEVERAL_NEGATIVE)
+            await state.clear()
+    elif deletion_type['deletion_type'] == 'delete_one':
+        if message.text in text_to_repeat:
             await delete_one(message)
             await message.answer(Russian.DELETE_ONE_POSITIVE)
             await state.clear()
-        except:
+        else:
             await message.answer(Russian.DELETE_ONE_NEGATIVE)
             await state.clear()
